requirements.txt
requests>=2.31.0
jinja2>=3.1.2
weasyprint>=60.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
